Disable setting of max_allowed_packet before analysis
As we run the analysis on a non-self-maintained DB server now, we are
not allowed to set the 'max_allowed_packet' option in MySQL now. The
option has been set by the DB administrators to an appropriate value,
but the Codeface command fails with an error due to missing privileges.

IMPORTANT: This is NOT wanted in upstream Codeface.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>